@@ -6,7 +6,7 @@
     "@types/jest": "^26.0.15",
     "@types/qs": "^6.9.7",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
+    "@types/react": "^18.0.12",
     "@types/react-dom": "^17.0.0",
     "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.7",
@@ -15,19 +15,19 @@
     "@testing-library/user-event": "^12.1.10"
   },
   "dependencies": {
-    "@mui/material": "^5.2.7",
-    "@mui/icons-material": "^5.2.5",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@emotion/css": "^11.7.1",
+    "@mui/material": "^5.8.4",
+    "@mui/icons-material": "^5.8.4",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@emotion/css": "^11.9.0",
     "@smartb/g2": "1.0.0-alpha.21",
     "connected-react-router": "^6.9.1",
     "qs": "^6.10.3",
     "deox": "^3.3.1",
     "history": "^4.10.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-app-rewired": "^2.1.11",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.4",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
