@@ -10,25 +10,24 @@
   "npmClient": "yarn",
   "useWorkspaces": true,
   "devDependencies": {
-    "@mui/material": "^5.2.7",
-    "@emotion/react": "^11.7.1",
-    "@emotion/css": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@rollup/plugin-commonjs": "^21.0.1",
+    "@mui/material": "^5.8.4",
+    "@emotion/react": "^11.9.3",
+    "@emotion/css": "^11.9.0",
+    "@emotion/styled": "^11.9.3",
+    "@rollup/plugin-commonjs": "^22.0.0",
     "@rollup/plugin-image": "^2.1.1",
-    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-node-resolve": "^13.3.0",
     "rollup-plugin-postcss": "^4.0.2",
-    "postcss": "^8.4.5",
-    "@storybook/addon-actions": "^6.5.5",
-    "@storybook/addon-essentials": "^6.5.5",
-    "@storybook/addon-links": "^6.5.5",
-    "@storybook/addons": "^6.5.5",
-    "@storybook/node-logger": "^6.5.5",
-    "@storybook/react": "^6.5.5",
-    "@storybook/theming": "^6.5.5",
+    "postcss": "^8.4.14",
+    "@storybook/addon-actions": "^6.5.9",
+    "@storybook/addon-essentials": "^6.5.9",
+    "@storybook/addon-links": "^6.5.9",
+    "@storybook/addons": "^6.5.9",
+    "@storybook/node-logger": "^6.5.9",
+    "@storybook/react": "^6.5.9",
+    "@storybook/theming": "^6.5.9",
     "@svgr/rollup": "^6.2.1",
     "@svgr/webpack": "5.5.0",
-    "@types/jest": "^25.1.4",
     "@typescript-eslint/eslint-plugin": "^2.26.0",
     "@typescript-eslint/parser": "^2.26.0",
     "babel-eslint": "^10.0.3",
@@ -46,17 +45,17 @@
     "husky": "^4.2.5",
     "lerna": "^4.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.0.4",
-    "pretty-quick": "^2.0.1",
+    "prettier": "^2.7.0",
+    "pretty-quick": "^3.1.3",
     "raw-loader": "^4.0.2",
-    "react-is": "^17.0.2",
-    "rollup": "^2.67.0",
+    "react-is": "^18.2.0",
+    "rollup": "^2.75.6",
     "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-typescript2": "^0.31.2",
-    "storybook-addon-designs": "^6.2.1",
-    "typescript": "^4.5.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "rollup-plugin-typescript2": "^0.32.1",
+    "storybook-addon-designs": "^6.3.1",
+    "typescript": "^4.7.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "husky": {
     "hooks": {
@@ -64,7 +63,7 @@
     }
   },
   "scripts": {
-    "storybook": "STORYBOOK_DISABLE_TELEMETRY=1 start-storybook -p 6006",
+    "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "chromatic": "npx chromatic",
     "publishWorkspaces:version": "lerna version ${VERSION} --no-git-tag-version --yes",
