@@ -27,20 +27,20 @@
     "dist"
   ],
   "peerDependencies": {
-    "@mui/material": "^5.2.7",
-    "@types/react": "^17.0.14",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@mui/material": "^5.8.4",
+    "@types/react": "^18.0.12",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@mui/icons-material": "^5.2.5",
+    "@mui/icons-material": "^5.8.4",
     "@smartb/g2-components": "^1.0.0-alpha.37",
     "@smartb/g2-forms": "^1.0.0-alpha.37",
     "@smartb/g2-layout": "^1.0.0-alpha.37",
     "@smartb/g2-notifications": "^1.0.0-alpha.37",
     "@smartb/g2-providers": "^1.0.0-alpha.37",
     "@smartb/g2-themes": "^1.0.0-alpha.37",
-    "react-query": "^3.38.1"
+    "react-query": "^3.39.1"
   },
   "publishConfig": {
     "@smartb:registry": "https://gitlab.smartb.city/api/v4/projects/52/packages/npm/"
