@@ -27,19 +27,19 @@
     "@smartb/g2-documentation": "^1.0.0-alpha.37"
   },
   "peerDependencies": {
-    "@mui/material": "^5.2.7",
-    "@types/react": "^17.0.14",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@mui/material": "^5.8.4",
+    "@types/react": "^18.0.12",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@mui/icons-material": "^5.2.5",
-    "@mui/lab": "^5.0.0-alpha.63",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/lab": "5.0.0-alpha.86",
     "@smartb/g2-components": "^1.0.0-alpha.37",
     "@smartb/g2-forms": "^1.0.0-alpha.37",
     "@smartb/g2-themes": "^1.0.0-alpha.37",
-    "@types/react-table": "^7.7.9",
-    "react-table": "^7.7.0"
+    "@types/react-table": "^7.7.12",
+    "react-table": "^7.8.0"
   },
   "files": [
     "dist"
