@@ -31,13 +31,13 @@
     "@smartb/g2-forms": "^1.0.0-alpha.37"
   },
   "peerDependencies": {
-    "@mui/material": "^5.2.7",
-    "@types/react": "^17.0.14",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@mui/material": "^5.8.4",
+    "@types/react": "^18.0.12",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@mui/icons-material": "^5.2.5",
+    "@mui/icons-material": "^5.8.4",
     "@smartb/g2-themes": "^1.0.0-alpha.37",
     "notistack": "3.0.0-alpha.1"
   },
