@@ -1,13 +1,13 @@
 import { MergeMuiElementProps } from '@smartb/g2-themes'
 import React from 'react'
-import { Route, Switch, SwitchProps } from 'react-router'
+import { Route, Routes, RoutesProps } from 'react-router-dom'
 import { NoMatchPage, NoMatchPageProps } from './NoMatchPage'
 
 interface RouterBasicProps {
   /**
    * The component display by the root if the user is authorized and if the current route match the one you gave
    */
-  children?: React.ReactNode
+  children?: React.ReactElement | React.ReactElement[]
   /**
    * By default the noMatch redirection goes to the `NoMatchPage` component but you can override it with this prop
    *
@@ -20,16 +20,17 @@ interface RouterBasicProps {
   noMatchPageProps?: NoMatchPageProps
 }
 
-export type RouterProps = MergeMuiElementProps<SwitchProps, RouterBasicProps>
+export type RouterProps = MergeMuiElementProps<RoutesProps, RouterBasicProps>
 
 export const Router = (props: RouterProps) => {
   const { children, noMatchComponent, noMatchPageProps, ...other } = props
   return (
-    <Switch {...other}>
+    <Routes {...other}>
       {children}
-      <Route path='*'>
-        {noMatchComponent || <NoMatchPage {...noMatchPageProps} />}
-      </Route>
-    </Switch>
+      <Route
+        path='*'
+        element={noMatchComponent || <NoMatchPage {...noMatchPageProps} />}
+      />
+    </Routes>
   )
 }
